Posts: Move editor post edit state to posts subtree
