+++ client/state/posts/reducer.js
@@ -8,6 +8,7 @@ import keyBy from 'lodash/keyBy';
  * Internal dependencies
  */
 import {
+	POST_EDIT,
 	POST_REQUEST,
 	POST_REQUEST_SUCCESS,
 	POST_REQUEST_FAILURE,
@@ -158,10 +159,39 @@ export function queriesLastPage( state = {}, action ) {
 	return state;
 }
 
+/**
+ * Returns the updated editor posts state after an action has been dispatched.
+ * The state maps site ID, post ID pairing to an object containing revisions
+ * for the post.
+ *
+ * @param  {Object} state  Current state
+ * @param  {Object} action Action payload
+ * @return {Object}        Updated state
+ */
+export function edits( state = {}, action ) {
+	switch ( action.type ) {
+		case POST_EDIT:
+			const { post, siteId, postId = '' } = action;
+			state = Object.assign( {}, state, {
+				[ siteId ]: Object.assign( {}, state[ siteId ] )
+			} );
+
+			state[ siteId ][ postId ] = Object.assign( {}, state[ siteId ][ postId ], post );
+			return state;
+
+		case SERIALIZE:
+		case DESERIALIZE:
+			return {};
+	}
+
+	return state;
+}
+
 export default combineReducers( {
 	items,
 	siteRequests,
 	queryRequests,
 	queries,
-	queriesLastPage
+	queriesLastPage,
+	edits
 } );
